+++ pytorch_forecasting/metrics.py
@@ -112,7 +112,7 @@ class Metric(LightningMetric):
         elif y_pred.ndim == 3:
             if y_pred.size(2) > 1:  # single dimension means all quantiles are the same
                 assert quantiles is not None, "quantiles are not defined"
-                y_pred = torch.quantile(y_pred, torch.tensor(quantiles), dim=2).permute(1, 2, 0)
+                y_pred = torch.quantile(y_pred, torch.tensor(quantiles, device=y_pred.device), dim=2).permute(1, 2, 0)
             return y_pred
         else:
             raise ValueError(f"prediction has 1 or more than 3 dimensions: {y_pred.ndim}")
@@ -551,7 +551,7 @@ class MultiHorizonMetric(Metric):
         else:
             losses = losses.sum()
             if not torch.isfinite(losses):
-                losses = torch.tensor(1e9)
+                losses = torch.tensor(1e9, device=losses.device)
                 warnings.warn("Loss is not finite. Resetting it to 1e9")
             self.losses = self.losses + losses
             self.lengths = self.lengths + lengths.sum()
@@ -1012,10 +1012,10 @@ class DistributionLoss(MultiHorizonMetric):
             quantiles = self.quantiles
         try:
             distribution = self.map_x_to_distribution(y_pred)
-            quantiles = distribution.icdf(torch.tensor(quantiles)[:, None, None]).permute(1, 2, 0)
+            quantiles = distribution.icdf(torch.tensor(quantiles, device=y_pred.device)[:, None, None]).permute(1, 2, 0)
         except NotImplementedError:  # resort to derive quantiles empirically
             samples = torch.sort(self.sample(y_pred, n_samples), -1).values
-            quantiles = torch.quantile(samples, torch.tensor(quantiles), dim=2).permute(1, 2, 0)
+            quantiles = torch.quantile(samples, torch.tensor(quantiles, device=samples.device), dim=2).permute(1, 2, 0)
         return quantiles